feat: move changes to menuPopup
