--- packages/components/src/Filter/FilterMultiSelect/subcomponents/MenuPopup/MenuPopover.module.scss
@@ -1,28 +1,0 @@
-@import '~@kaizen/design-tokens/sass/spacing';
-@import '~@kaizen/design-tokens/sass/shadow';
-@import '~@kaizen/design-tokens/sass/border';
-@import '~@kaizen/design-tokens/sass/color';
-
-// figma hard coded: https://www.figma.com/file/eZKEE5kXbEMY3lx84oz8iN/%E2%9D%A4%EF%B8%8F-UI-Kit%3A-Heart?node-id=22814%3A96966
-$menu-container-width: 294px;
-$menu-container-max-height: 312px;
-
-.menuPopover {
-  z-index: 1000; // from $ca-z-index-dropdown
-  box-sizing: border-box;
-  background: $color-white;
-  color: $color-purple-800;
-  border-radius: $border-solid-border-radius;
-  box-shadow: $shadow-large-box-shadow;
-  padding: $spacing-sm 0;
-  margin-top: $spacing-xs;
-  text-align: start;
-  width: $menu-container-width;
-  
-}
-
-.menuPopover[popover]:popover-open {
-  z-index: unset;
-  margin: 0;
-  inset: unset;
-}
